updates in argument parser & progress bar in download

+++ daemon.py
@@ -4,16 +4,31 @@ import json
 from pymongo import MongoClient, GEO2D
 import os
 import argparse
+import sys
 
 # downloads csv.gz file from opencellid
 def downloadCsv(transactions_source_csv_gz, token):
 	# open in binary mode
 	with open(transactions_source_csv_gz, "wb") as file:
+		print('Downloading ' + transactions_source_csv_gz + '.')
 		# get request
 		request = get('https://download.unwiredlabs.com/ocid/downloads?token=' + token + '&file=cell_towers.csv.gz', stream=True)
-		for chunk in request.iter_content(chunk_size = 1024):
-			if chunk:
+		# request = get('http://ipv4.download.thinkbroadband.com/100MB.zip', stream=True)
+		total_length = request.headers.get('content-length')
+		if total_length is None: # no content length header
+			print('Could not download ' + transactions_source_csv_gz +  '. Maybe your access token is limited.')
+			return False
+		else:
+			dl = 0
+			total_length = int(total_length)
+			for chunk in request.iter_content(chunk_size = 4096):
+				dl += len(chunk)
 				file.write(chunk)
+				done = int(50 * dl / total_length)
+				sys.stdout.write("\r[%s%s]" % ('=' * done, ' ' * (50-done)) )    
+				sys.stdout.flush()
+			return True
+
 
 # updates database in order to store latitude and longitude as parts of "coords"
 # also ensures "coords" is a 2D index (this way we can easily find neighbour cells using MongoDB functions)
@@ -32,7 +47,6 @@ def toMongo(dest, data, drop, idxs=[]):
 	if (drop):
 		dest.drop()
 	for chunk in data:
-		print(chunk)
 		dest.insert(json.loads(chunk.to_json(orient='records')))
 	updateLoc(dest)
 	for idx in idxs:
@@ -43,26 +57,19 @@ def updateDatabase(db_host, db_name, db_port, chunk_size, drop, transactions_col
 	# establish connection:
 	mongoClient = MongoClient(db_host, db_port)
 	mongoDb = mongoClient[db_name]
-
-	try:
-		# load data:
-		transactions = pd.read_csv(
-			transactions_source_csv_gz,
-			compression='gzip',
-			chunksize=chunk_size)
-		# insert data:
-		toMongo(mongoDb[transactions_collection],
-				 transactions,
-				 drop,
-				 ['mcc','net','area','cell'])
-		# close connection:
-		mongoClient.close()
-	except:
-		print("Unexpected error. Maybe your access token is limited.")
+	print()
+	print('Updating ' + db_host + ":" + str(db_port) + ":" + db_name + ":" + transactions_collection + ' using dataset ' + transactions_source_csv_gz + '.')
+	# load data:
+	transactions = pd.read_csv(transactions_source_csv_gz, compression='gzip', chunksize=chunk_size)
+	# insert data:
+	toMongo(mongoDb[transactions_collection], transactions, drop, ['mcc','net','area','cell'])
+	# close connection:
+	mongoClient.close()
 
 def main():
 	parser = argparse.ArgumentParser(description='Download open cell id dataset and store it to a MongoDB database.')
-	parser.add_argument('-tk', '--token', default='9d9a0a15012fc1', help='Access token for unwiredlabs open cell id.')
+	# open cell id tokens: 9d9a0a15012fc1, 92af25912f69f6
+	parser.add_argument('-tk', '--token', default='92af25912f69f6', help='Access token for unwiredlabs open cell id.')
 	parser.add_argument('-ht', '--db_host', default='localhost', help='MongoDB host address.')
 	parser.add_argument('-n', '--db_name', default='open_cell_id', help='MongoDB name.')
 	parser.add_argument('-p', '--db_port', type=int, default=27017, help='MongoDB port.')
@@ -70,14 +77,21 @@ def main():
 	parser.add_argument('-d', '--drop', type=bool, default=False, help='Drops MongoDB collection if set to True.')
 	parser.add_argument('-tc', '--transactions_collection', default='cell_towers', help='Name of transactions collection.')
 	parser.add_argument('-ts', '--transactions_source_csv_gz', default='cell_towers.csv.gz', help='Name of transactions source.')
+	parser.add_argument('-dl', '--download_transactions_source_csv_gz', type=bool, default=False, help='If set to True transactions_source_csv_gz is downloaded even if it already exists.')
+	parser.add_argument('-r', '--remove_transactions_source_csv_gz', type=bool, default=False, help='Removes transactions_source_csv_gz if set to True.')
 
 	args = parser.parse_args()
+	
 
-	print('Downloading data...')
-	downloadCsv(args.transactions_source_csv_gz, args.token)
-	print('Updating database...')
-	updateDatabase(args.db_host, args.db_name, args.db_port, args.chunk_size, args.drop, args.transactions_collection, args.transactions_source_csv_gz)
-	os.remove(args.transactions_source_csv_gz)
+	if args.download_transactions_source_csv_gz:
+		if downloadCsv(args.transactions_source_csv_gz, args.token):
+			updateDatabase(args.db_host, args.db_name, args.db_port, args.chunk_size, args.drop, args.transactions_collection, args.transactions_source_csv_gz)
+	elif os.path.isfile(args.transactions_source_csv_gz):
+		updateDatabase(args.db_host, args.db_name, args.db_port, args.chunk_size, args.drop, args.transactions_collection, args.transactions_source_csv_gz)
+	else:
+		print(args.transactions_source_csv_gz + ' file not found.')
+	if args.remove_transactions_source_csv_gz:
+		os.remove(args.transactions_source_csv_gz)
 
 if __name__ == '__main__':
 	main()
